Add comments on why sections are modified for performance
ES6, object/array spreading and the like make the code much more readable and easier 
to reason about.

However, this trade-off comes at the cost of performance so it would be nice to future 
devs to show why things are written the way they are and to not try and refactor them
into something more readable (and less performant).

Normally, readable code > super optimized code but selectors can be called very often and the overhead associated with those calls has to be minimal.